Handled multithreaded locking and lifetime issues with external sources page

+++ src/pages/page_cover.h
@@ -22,7 +22,7 @@ private:
     int mCurrentMusicBeat;
 
     uint32_t musicInstanceID;
-    uint32_t audioObjectID;
+    uint64_t audioObjectID;
 
     void Start() override;
